+++ src/main/java/com/bmaster/StreamingWebUI/Controller/RootController.java
@@ -11,28 +11,9 @@ public class RootController {
 	@RequestMapping(path = "/", method = RequestMethod.GET)
 	public String index(Model model) {
 		//model.addAttribute("UserId", "Ghostranger");
-		return "login.html";
-	}
-	
-	
-	@RequestMapping(path="/login", method=RequestMethod.POST)
-	public String login(Model model) {
-		model.addAttribute("sidepane","dashboard");
-		return "redirect:/dashboard";
-	}
-	
-	@RequestMapping(path="/dashboard", method=RequestMethod.GET)
-	public String dashboard(Model model) {
-		model.addAttribute("sidepane","dashboard");
-		model.addAttribute("ab","yes");
 		return "dashboard/index.html";
 	}
-	@RequestMapping(path="/kafka",method=RequestMethod.GET)
-	public String kafkaTrigger(Model model) {
-		model.addAttribute("sidepane","kafka");
-		//model.addAttribute("ab","yes");
-		return "kafka/index.html";
-	}
+	
 	
 	@RequestMapping(path="/jms",method=RequestMethod.GET)
 	public String jmsTrigger(Model model) {
@@ -40,11 +21,6 @@ public class RootController {
 		return "jms/index.html";
 	}
 	
-	@RequestMapping(path="/adls",method=RequestMethod.GET)
-	public String adlsTrigger(Model model) {
-		model.addAttribute("sidepane","adls");
-		return "adls/index.html";
-	}
 	
 	@RequestMapping(path="/kinesis",method=RequestMethod.GET)
 	public String kinesisTrigger(Model model) {
@@ -52,10 +28,6 @@ public class RootController {
 		return "kinesis/index.html";
 	}
 
-	@RequestMapping(path="/hdfs",method=RequestMethod.GET)
-	public String hdfsTrigger(Model model) {
-		model.addAttribute("sidepane","hdfs");
-		return "hdfs/index.html";
-	}
+	
 	
 }
