+++ src/main/java/com/bmaster/StreamingWebUI/Controller/hdfs/HdfsController.java
@@ -1,0 +1,16 @@
+package com.bmaster.StreamingWebUI.Controller.hdfs;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+
+@Controller
+@RequestMapping(path="/hdfs")
+public class HdfsController {
+	@RequestMapping(path="",method=RequestMethod.GET)
+	public String hdfsTrigger(Model model) {
+		model.addAttribute("sidepane","hdfs");
+		return "hdfs/index.html";
+	}
+}
