+++ src/main/resources/templates/error.html
@@ -3,7 +3,7 @@
     lang="en"
     xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
     xmlns:sec="http://www.thymeleaf.org/thymeleaf-extras-springsecurity4"
-    xmlns:th="http://www.thymeleaf.org">
+    xmlns:th="http://www.thymeleaf.org" style="background-color:#f8f9fc">
 <head>
   <meta charset="utf-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
